rework recent link component
